remove compatibility remnants we no longer wish to support
There are a bunch of things that we don't want to keep around, and
now we've made version 4 (not used by Juju yet, and incompatible
for other reasons) we are free to delete things that are no longer
used or that are easily replaced by other functionality in the package.

Specifically:

- LocalRepository is a throwback to the times when all local charms were
stored in $CHARMDIR/$SERIES/$CHARMNAME.
- The global CacheDir is problematic and unnecessary. We delete it instead
of adding more complexity (c.f. PR #110) We change the charm and
bundle interface to explicitly take a file name to write to instead of using
CacheDir. With LocalRepository gone, we can change charmrepo.Interface so
it returns a *CharmArchive instead of the more general charm interface.
- CharmRevision does not belong in the params package, as it's not part
of the charmstore API contract. Also, nothing in Juju uses the fact that
it returns the SHA256 checksum of the charms, so we move the type to csclient
and remove the Sha256 field.
- The legacy charm store is no longer running, so there's no need to keep
client support for it.