csclient: add DockerResourceDownloadInfo method
This provides an entry point to the docker resource download
endpoint. We want it to fail fast if it's used to download a
non-docker resource, so we change the response parsing
code to check the content type before parsing the content.

As the httprequest package already provides that functionality
(and we're already dependent on that), we use that.