Add charmstore method to resolve charm/bundles with a preferred channel
Prior to this change, the client could enforce a particular channel for
*all* lookups. However, we the recent charm.v6 changes we are now able
to specify a per-application channel. As a result, we need a new resolve
method whose signature allows us to specify the preferred channel to
use.

If no channel is specified, the clients forced channel (if provided)
will be used as before and then the lookup code will fall back to
selecting the best (most stable) channel version based on the metadata
obtained by the store.