Update the charm repo.
The brings in the new charm changes around error messages. The sad
reality of the code though is that we now have 3 different error types
that we have to negotiate between...

 - std error
 - gopkg.in/errgo
 - github.com/juju/errors

Considering this is infact a juju project, we should drop the errgo one
and be done with it. The code path itself has to marshal and unmarshal
the errors from juju/errors -> errgo -> juju/errors and each one has
different symantics. Not to mention we actually loose the stack trace
from the marshalling and unmarshalling, which is a damn shame.